move blob function to blob file
